Create GlVectorUtil.cpp
