+++ ProjOpenGl/main.cpp
@@ -405,31 +405,6 @@ void setupTextures() {
   engineExhaustTexture.createOneColor(exhaustRgb[0], exhaustRgb[1], exhaustRgb[2]);
 }
 
-// TODO: move to some file in UtilitiesGl
-// Vectors originalDirection and destinationDirection must be normalized.
-void getDirectionDifference(
-    const glm::dvec3 & originalDirection, const glm::dvec3 & destinationDirection,
-    double & rotationAngleDeg, glm::dvec3 & rotationAxis)
-{
-  if(originalDirection == destinationDirection) {
-    rotationAxis = glm::dvec3(0, 1, 0);
-    rotationAngleDeg = 0;
-    return;
-  }
-
-  if(originalDirection == -destinationDirection) {
-    rotationAxis = glm::dvec3(0, 1, 0);
-    rotationAngleDeg = 180;
-    return;
-  }
-
-  rotationAxis = glm::cross(originalDirection, destinationDirection);
-
-  auto dotProduct = glm::dot(originalDirection, destinationDirection);
-  auto rotationAngleRad = acos(dotProduct);
-  rotationAngleDeg = rotationAngleRad / M_PI * 180.0;
-}
-
 void drawSpacecraft(const Spacecraft & spacecraft, double size,
                     const Color & spacecraftColor, const Texture & spacecraftTexture,
                     const Color & exhaustColor, const Texture & exhaustTexture)
