--- UtilitiesGl/GlVectorConversion.h
@@ -1,4 +1,0 @@
-#include <glm/glm.hpp>
-#include "Vector.h"
-
-glm::dvec3 vectorToGlmDvec3(const Vector & v);
