+++ proj-opengl/bixi.pro
@@ -33,7 +33,7 @@ HEADERS += ../PhysicalObjects/PhysicalObject.h \
            ../Utilities/ScopedFile.h \
            ../UtilitiesGl/UtilGlDraw.h \
            ../UtilitiesGl/Texture.h \
-           ../UtilitiesGl/GlVectorConversion.h \
+           ../UtilitiesGl/GlVectorUtil.h \
            ../ProjOpenGl/main.h \
            ../ProjOpenGl/Camera.h \
            ../ProjOpenGl/Keyboard.h \
@@ -53,7 +53,7 @@ SOURCES += ../PhysicalObjects/PhysicalObject.cpp \
            ../Utilities/ImageData.cpp \
            ../UtilitiesGl/UtilGlDraw.cpp \
            ../UtilitiesGl/Texture.cpp \
-           ../UtilitiesGl/GlVectorConversion.cpp \
+           ../UtilitiesGl/GlVectorUtil.cpp \
            ../ProjOpenGl/main.cpp \
            ../ProjOpenGl/Camera.cpp \
            ../ProjOpenGl/Keyboard.cpp \
